+++ Changelog.md
@@ -9,11 +9,17 @@ Types of changes: New, Improved, Fixed, Removed.
 
 ## [Unreleased]
 
+## [1.0.0]
+
 - New: Added zed extension
 - New: Abc command has new flag `--printed` to direct output to the terminal instead of a file
 - New: Added `dense` field label for tune headers
 - Improved: CLI output flag (`-o`) is now optional for abc and pdf commands
 - Improved: Uses strict concurrency for swift 6 language mode
+
+### Breaking changes
+
+Ommitting the `-o` command line option will no longer print the output. To print the output, use `--printed`.
 
 ## [0.4.1] - 14 Feb 2025